+++ mmedit/utils/__init__.py
@@ -1,6 +1,6 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 from .cli import modify_args
-from .img_utils import reorder_image, tensor2img, to_numpy
+from .img_utils import get_box_info, reorder_image, tensor2img, to_numpy
 from .io_utils import MMEDIT_CACHE_DIR, download_from_url
 # TODO replace with engine's API
 from .logger import print_colored_log
@@ -17,5 +17,5 @@ __all__ = [
     'download_from_url', 'get_sampler', 'tensor2img', 'random_choose_unknown',
     'add_gaussian_noise', 'adjust_gamma', 'make_coord', 'bbox2mask',
     'brush_stroke_mask', 'get_irregular_mask', 'random_bbox', 'reorder_image',
-    'to_numpy'
+    'to_numpy', 'get_box_info'
 ]
